pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sqlgenx"
version = "0.1.0"
description = "AI-powered CLI for Smart SQL Generation"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.6",
    "langgraph>=0.0.20",
    "chromadb>=0.4.0",
    "sqlglot>=20.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
sqlgenx = "sqlgenx.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["sqlgenx*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true